requirements.txt
flask
syncedlyrics
google-generativeai
gunicorn
firebase-admin